Move ObjectCache for FFI into an internal module
ObjectCache  for the native implementation moved into an internal
module by f824f6d0d.

I think for FFI needs to move too because the test for ObjectCache
expects to exist constant values under an internal module.
https://github.com/protocolbuffers/protobuf/blob/610574c480a124cf1636cbe7e29cc42207e6559e/ruby/tests/object_cache_test.rb#L7

Without this change, that test fails with `NameError`.
