+++ ruby/lib/google/protobuf/ffi/repeated_field.rb
@@ -367,7 +367,7 @@ module Google
 
         # Should always be the last expression of the initializer to avoid
         # leaking references to this object before construction is complete.
-        OBJECT_CACHE.try_add(@array.address, self)
+        Google::Protobuf::Internal::OBJECT_CACHE.try_add(@array.address, self)
       end
 
       ##
