+++ ruby/lib/google/protobuf_ffi.rb
@@ -11,6 +11,7 @@ require 'google/protobuf/ffi/internal/type_safety'
 require 'google/protobuf/ffi/internal/pointer_helper'
 require 'google/protobuf/ffi/internal/arena'
 require 'google/protobuf/ffi/internal/convert'
+require 'google/protobuf/ffi/internal/object_cache'
 require 'google/protobuf/ffi/descriptor'
 require 'google/protobuf/ffi/enum_descriptor'
 require 'google/protobuf/ffi/field_descriptor'
@@ -20,7 +21,6 @@ require 'google/protobuf/ffi/service_descriptor'
 require 'google/protobuf/ffi/descriptor_pool'
 require 'google/protobuf/ffi/file_descriptor'
 require 'google/protobuf/ffi/map'
-require 'google/protobuf/ffi/object_cache'
 require 'google/protobuf/ffi/repeated_field'
 require 'google/protobuf/ffi/message'
